refactor(orchestrator): remove dead cycle funcs, split Run, drop test-only MarkRendered
